refactor slider loader
