+++ widgets/slider/layout.hbs
@@ -3,6 +3,7 @@
     <div class="tiles-scroll-button" id="scroll-up">
       <span class="widget-icon chevron-left"></span>
     </div>
+    <div class="slider-loading loading"></div>
     <div class="ugc-tiles">
       {{#tiles}}
       {{>tpl-tile options=../options.widgetOptions.config}}
